Add basic pip install capability, reorganised directory structure
Pip install capability installs from requirements file.
