+++ PK_modelling.egg-info/PKG-INFO
@@ -1,10 +1,82 @@
-Metadata-Version: 1.0
+Metadata-Version: 1.1
 Name: PK-modelling
 Version: 0.1.3
-Summary: UNKNOWN
-Home-page: UNKNOWN
-Author: UNKNOWN
-Author-email: UNKNOWN
-License: UNKNOWN
-Description: UNKNOWN
+Summary: 'Library for producing, solving and visualising pharmacokinetic models.'
+Home-page: https://github.com/Software-Engineering-BBSRC-Group-6/PK_modelling
+Author: Joe Pollacco
+Author-email: joseph.pollacco@dtc.ox.ac.uk
+License: MIT
+Description: [![Python application](https://github.com/Software-Engineering-BBSRC-Group-6/PK_modelling/actions/workflows/run-unittests.yml/badge.svg)](https://github.com/Software-Engineering-BBSRC-Group-6/PK_modelling/actions/workflows/run-unittests.yml)
+        [![Check systems capacity](https://github.com/Software-Engineering-BBSRC-Group-6/PK_modelling/actions/workflows/check-systems-compat.yml/badge.svg)](https://github.com/Software-Engineering-BBSRC-Group-6/PK_modelling/actions/workflows/check-systems-compat.yml)
+        [![codecov](https://codecov.io/gh/Software-Engineering-BBSRC-Group-6/PK_modelling/branch/main/graph/badge.svg?token=gdzMuuonBd)](https://codecov.io/gh/Software-Engineering-BBSRC-Group-6/PK_modelling)
+        [![Documentation Status](https://readthedocs.org/projects/pk-proj/badge/?version=latest)](https://pk-proj.readthedocs.io/en/latest/?badge=latest)
+        # Pharmacokinetic Modelling Group Project
+        
+        A PharmacoKinetic (PK) modelling function for analysis of injected solute dynamics over time, developed by Group 8 of the 2021 Software Engineering course. This model has been developed with the end-user in mind, and includes an easy-to-use interface to guide refinement of tissue models.
+        # Authors
+        
+        Amit Halkhoree - amit.halkhoree@dtc.ox.ac.uk \
+        Cameron Anderson - cameron.anderson@dtc.ox.ac.uk \
+        Dan Hudson - alexander.hudson@dtc.ox.ac.uk \
+        Ishaan Kapoor - ishaan.kapoor@dtc.ox.ac.uk \
+        Joseph Pollacco - joseph.pollacco@dtc.ox.ac.uk \
+        Samuel Johnson - samuel.johnson@dtc.ox.ac.uk
+        
+        # Background
+        Ref: https://sabs-r3.github.io/software-engineering-projects/01-introduction/index.html.
+        
+        The field of Pharmacokinetics (PK) provides a quantitative basis for describing the delivery of a drug to a patient, the diffusion of that drug through the plasma/body tissue, and the subsequent clearance of the drug from the patient’s system. PK is used to ensure that there is sufficient concentration of the drug to maintain the required efficacy of the drug, while ensuring that the concentration levels remain below the toxic threshold. Pharmacokinetic (PK) models are often combined with Pharmacodynamic (PD) models, which model the positive effects of the drug, such as the binding of a drug to the biological target, and/or undesirable side effects, to form a full PKPD model of the drug-body interaction. This project will only focus on PK, neglecting the interaction with a PD model.
+        
+        # Model overview
+        
+        Our model consists of three principal modules:
+        - A solver (solver.py) that computes the distribution of solute in different tissue compartments over time, given input parameters including the number of compartments to model, dose,  mode of administration  (intravenous or subcutaneous), compartment volume and transition rate constants.
+        - A user interface (ux.py) that prompts the user to configure the model
+        - A visualisation module (XXX.py) that generates and saves comparative plots to facilitate evaluation of PK using different administration methods.
+        
+        # Folder structure
+        
+        This repository includes the following files/folders:
+        
+        - __pycache__: directory including bytecode cache generated by Python
+        - .github/workflows: directory of github actions ensuring continuous integration (CI) of repository updates 
+        - .pytest/cache: directory including bytecode cache generated by PyTest, which performs unit tests on key executables
+        - .vscode: virtualstudio code workspace for development purposes
+        - docs: files required for automated document production
+        - README.md: text file populating this guide
+        - tests: unit tests fed to pytest for continuous integration
+        - venv: virtual environment including dependencies for this package
+        - .gitignore: file controlling which files are/are not updated by git during development
+        - definitions.py: Key helper functions for solver.py
+        - requirements.txt: Dependencies required for proper package functioning
+        - solver.py: Solver model for PK modelling
+        - ux.py: User interface executable
+        
+        
+        
+        
+        MIT License
+        
+        Copyright (c) 2021 Software-Engineering-BBSRC-Group-6
+        
+        Permission is hereby granted, free of charge, to any person obtaining a copy
+        of this software and associated documentation files (the "Software"), to deal
+        in the Software without restriction, including without limitation the rights
+        to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+        copies of the Software, and to permit persons to whom the Software is
+        furnished to do so, subject to the following conditions:
+        
+        The above copyright notice and this permission notice shall be included in all
+        copies or substantial portions of the Software.
+        
+        THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+        IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+        FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+        AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+        LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+        OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+        SOFTWARE.
+        
+Keywords: pharmokinetic,pharmacokinetic,compartment,ODE
 Platform: UNKNOWN
+Classifier: Programming Language :: Python :: 3.9
