+++ PK_modelling.egg-info/SOURCES.txt
@@ -1,7 +1,16 @@
 README.md
+pk_model_main_pipeline.py
 setup.cfg
 setup.py
 PK_modelling.egg-info/PKG-INFO
 PK_modelling.egg-info/SOURCES.txt
 PK_modelling.egg-info/dependency_links.txt
+PK_modelling.egg-info/not-zip-safe
+PK_modelling.egg-info/requires.txt
 PK_modelling.egg-info/top_level.txt
+pk_model/__init__.py
+pk_model/definitions.py
+pk_model/dosing.py
+pk_model/solver.py
+pk_model/ux.py
+pk_model/visualiser.py
