+++ pk_model/ux.py
@@ -4,6 +4,7 @@
 # and then returns a json with the answers to be solved
 import json
 import time
+import os
 
 
 def user_input():
@@ -246,13 +247,12 @@ def param_to_file():
     Writes dictionary generated from user_input() and generates .json file
     """
     data = user_input()
-    fname = data.get('curr_datetime')
-    f = open(fname+".json", "w")
+    fname = 'jsons/' + data.get('curr_datetime') + '.json'
+    f = open(fname , 'w')
     json.dump(data, f)
     f.close()
-    filename = fname+".json"
 
-    return filename
+    return fname
 
 
 def num_models():
