+++ pk_model_main_pipeline.py
@@ -1,10 +1,10 @@
 # Helper script which runs the entire pipeline and handles user input,
 # solving models, visualising models
 
-from ux import num_models, param_to_file
-from solver import build_and_solve_model
-from dosing import build_dose
-from visualiser import make_plots
+from pk_model.ux import num_models, param_to_file
+from pk_model.solver import build_and_solve_model
+from pk_model.dosing import build_dose
+from pk_model.visualiser import make_plots
 
 # Get number of models that the user wants to compare.
 N = num_models()
@@ -13,7 +13,7 @@ filenames = []
 for i in range(N):
     # Get user input to build a model.
     #current_json = param_to_file()
-    current_json = '1634833362.6048021.json'
+    current_json = 'jsons/1634833362.6048021.json'
 
     # Transform that user input into a dosing protocol function.
     dosing_function = build_dose(current_json)
