+++ requirements.txt
@@ -21,6 +21,7 @@ pyparsing==2.4.7
 pytest==6.2.5
 python-dateutil==2.8.2
 requests==2.26.0
+sympy==1.6.2
 scipy==1.7.1
 six==1.16.0
 toml==0.10.2
