+++ setup.cfg
@@ -1,3 +1,21 @@
 [metadata]
 name = PK_modelling
 version = 0.1.3
+description = 'Library for producing, solving and visualising pharmacokinetic models.'
+url = https://github.com/Software-Engineering-BBSRC-Group-6/PK_modelling
+long_description = file: README.md, LICENSE
+author = 'Dan Hudson, Ishaan Kapoor, Sam Johnson, Amit Halkhoree, Cameron Anderson, Joe Pollacco'
+author_email = joseph.pollacco@dtc.ox.ac.uk
+keywords = pharmokinetic, pharmacokinetic, compartment, ODE
+license = MIT
+classifiers =
+    Programming Language :: Python :: 3.9
+
+[options]
+zip_safe = False
+include_package_data = False
+packages = pk_model
+scripts =
+    pk_model_main_pipeline.py
+    
+
